requirements.txt
torch==2.7.0
torchvision==0.22.0
numpy>=1.26.0
Pillow>=10.2.0
streamlit==1.28.0
scikit-learn==1.3.0
matplotlib==3.7.2
pandas==2.0.3
tqdm==4.65.0 